ci: run tests with and without PAPI debug enabled
Tests should make sure real use cases work as expected. Some tests might
not working correctly if -O0 is used as optimization level in the
compiler. For example, the ROCm runtime submits a kernel of 4 waves if
the tests are built using -O0, which makes the tests fail.

Update the github test configuration matrix to include testing without
PAPI debug.