configure: search for rocm_smi headers in PAPI_ROCMSMI_ROOT
The configure script used to search for rocm_smi headers in
PAPI_ROCM_ROOT instead of PAPI_ROCMSMI_ROOT. This was because the rocm
headers are typically installed under the same root. However, with
rocm-6.0.0 the rocm_smi.h causes a failure while building the sysdetect
component in PAPI (component that is enabled by default). Thus, we now
look explicitly for the rocm_smi header in PAPI_ROCMSMI_ROOT instead in
order to isolate the sysdetect & rocm components from rocm_smi.