Update:Use fgets in place of fscanf functions to avoid possible buffer overflows
There was a bug in the buffer overflow fix below.
Use fgets in place of fscanf functions to avoid possible buffer overflows
https://github.com/icl-utk-edu/papi/commit/ec2aa022fee2a1d0decf1d5b2e7e28a4ca2cf794

The buffer overflow fix rewrites fscanf() to fgets(),
but there is a difference in the specifications between fscanf() and fgets():
in fscanf(), "\n" is not read, but in fgets(), "\n" is also read,
so it seems necessary to remove the unnecessary "\n" to do the same.
For this reason, generic_get_memory_info() could not obtain cache information
strings from allocation_policy, type, and write_policy under
the /sys/devices/system/cpu/cpu0/cache/index[012] directory as expected.

The buffer overflow update fix allows us to do the same thing as fscanf()
by rewriting fscanf() to fgets()+sscanf().

Signed-off-by: Masahiko, Yamada <yamada.masahiko@fujitsu.com>