perf_event/libpfm4_events: add initial heterogenous CPU support
Alderlake and Raptorlake should now be able to properly enumerate
events for both Power and Efficiency cores on heterogeneous systems.

We're still testing if PAPI properly handles measurements where
applications end up running on both types of cores.

The detection of E-cores is currently a bit of a hack, we're going
to discuss with upstream libpfm4 to see if we can get a cleaner
interface for this.