cuda: When making the cuda tests do not assume nvcc on $PATH is one being used
A check in the Makefile for the cuda tests assumed the nvcc found on
$PATH and the one referred to by $(NVCC) were the same.  It is
possible to have multiple versions of cuda installed on the system and
the one being used for $(NVCC) may not be the same as the one found on
the default path.  Should always be use $(NVCC) in the Makefile to
ensure getting the same version of nvcc.