intel_gpu: properly reset metrics
Reset the appropriate fields in the context struct when PAPI_cleanup()
is called.
This corresponds to intel_gpu_update_control_state() being called with
the value 'count' equal to zero.

Also reset the internal metric counts when PAPI_start() is called, so
that the previous values do not persist after the next PAPI_start().

These changes have been tested on the Intel Ponte Vecchio architecture.