CUDA Tests: Conditionally adding -fpic to nvcc.
If the PAPI shared library is built (either because the user explicitly configured with --with-shared-lib=yes, or by default because the user did not specify anything) then nvcc will be instructed to create position-independent code through the flag "-fpic". If the user has disabled the shared library (by configuring with --with-shared-lib=no) then this flag is not passed to nvcc.