validation_tests: Add load/store ARM assembly testcode
The previous load/store validation tests were being optimized by
the compiler in a way that caused the tests to mispredict the
amount of memory instructions that are generated. This made it
appear like the counters were incorrect, when it was really the
test being inaccurate.

To fix this, add assembly testcode for ARM to eliminate the
problem of compiler optimizations. When load/store testcode
is unavailable for the current platform, default back to the
original matrix multiplication test.