sysdetect: allow entire cache level array to be used
The cpu_get_cache_info() function has a check that ensures the level of
cache info being acquired is within the bounds of the allowed number of
memory levels PAPI supports (PAPI_MAX_MEM_HIERARCHY_LEVELS). Previously
the check has used >= instead of >, preventing the last entry from being
used, causing the function to error out on the last level of the cache
on Intel Raptorlake. Change the check to allow the entire array to be
used.