sysdetect: eliminate infinite loop
Because the sysdetect component is designed to assume homogeneous cores,
heterogeneous CPUs like Intel Raptorlake can have the total number of
available threads be overestimated. The worst impact of this is an
infinite loop in get_thread_affinity() of linux_cpu_utils.c. Fix
infinite loop by adding a check to ensure /sys/ entry exists for the
target thread.