intel_gpu: remove unnecessary linker flags
Since the flag -lstdc++ is already included from the configure
script, it does not need to be included in the component's Rules
and Makefiles.
Also, -g is already included in the recipes, it is not needed in
the linker flags variable.

These changes have been tested on the Intel Max 1550 GPU.