topdown: Prevent segfault on heterogeneous CPUs
All of Intel's heterogeneous CPUs that support the PERF_METRICS MSR only
support it for their performance (p-core) cores. This means that if a
program that is being measured using the topdown component in PAPI
happens to be rescheduled to a e-core during its runtime, PAPI will
segfault.

To fix this, add a check in _topdown_start() and _topdown_stop() to exit
gracefully if the core affinity of the process has changed to an
unsupported core type.