topdown: relocated core type checks
To prevent programs using the topdown component on heterogeneous
processors that only supply the PERF_METRICS MSR on some of their cores
from segfaulting due to trying to read the MSR after being moved to an
unsupported core type, the topdown component periodically checks it is
on a supported core and exits if not.

Previously, this check occured at the start of PAPI_start and PAPI_stop.
After writing a script that starts a program being calipered with the
topdown component and moves it to an unsupported core after a random
amount of time, for N=100,000 tests the heterogeneous checks failed to
prevent a segmentation fault 0.08% of the time. This patch moves the
heterogeneous checks to occur only directly before the rdpmc calls,
resulting in cleaner code and a reduced segfault prevention failure
rate of 0.064%.

While it is frustrating that the failure rate is non-zero, since there
appears to be no way to tell a process to ignore changes to its
affinity, I believe there to be no perfect solution at this time.