Remove some preset events for FUJITSU-MONAKA
The following preset events of FUJITSU-MONAKA are not
counted properly:

PAPI_L3_DCM
PAPI_L3_TCM
PAPI_PRF_DM
PAPI_L3_DCH
PAPI_L3_TCH

Specifically, the native event that is the source of
the above preset event is always counted as a miss,
regardless of whether the L3 prefetch is a hit or a miss.
So I remove these events in papi_events.csv.