components: improper usage of PAPI_END macro
PAPI_END is a macro defined in papiStdEventDefs.h to denote the end of
the list of preset macros. However, it was being used as an error code
in various components in cases unrelated to the number of presets.

This commit changes this to a more appropriate error code: PAPI_ENOEVNT.

These changes have been tested with ROCm 6.3.1 on Frontier.