topdown: Use librseq to protect rdpmc on het cpus
On Intel's heterogeneous multicore processors such as Raptor Lake, the
PERF_METRICS MSR is only available on the performance cores (p-cores).
If the rdpmc instruction is executed attempting to access the MSR while
the process is on an efficient core (e-core), a segmentation fault
occurs.

Previously, the topdown component has used a simple check before every
execution of the rdpmc instruction to ensure the core the program is
bound to is a p-core. However, this can fail if the program is moved to
another core between the check and the execution of rdpmc. While rare, a
worst-case scenario test that repeatedly moves a program which is using
the topdown component from p-core to e-core at a random time saw 338
segmentation faults out of 1 million affinity switches (a 0.0338% error
rate). This is a non-zero number of segmentation faults, and we can do
better.

Use librseq to protect the rdpmc instruction with a restartable
sequence (rseq). When the process is preempted by an affinity change,
the sequence immediately aborts and can be restarted. By keeping the
check that the process is on a p-core and the rdpmc instruction itself
within the critical section of the rseq, it is guaranteed that the
rdpmc instruction will never be executed on an invalid core. The same
test described in the previous paragraph sees 0 segmentation faults.