Update libpfm4 Current with commit 86a9dc23f62ea14651dbe22b58211747f9f3d8fe Author: Stephane Eranian <eranian@gmail.com> Date:   Wed Oct 1 21:09:17 2025 -0700
    add support for ARM Cortex X4 core PMU

    Adds ARM Cortex X4 core PMU support.
    Based on:
                https://github.com/ARM-software/data/blob/master/pmu/cortex-x4.json

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

Note: The PAPI team currently does not have access to an ARM Cortex X4. However,
a sanity check was done on an ARM Neoverse V2 with the following PAPI utilities:
papi_avail, papi_component_avail, papi_native_avail, and papi_command_line. With all
the aforementioned utilities functioning as expected. Along with this, the output for
the script run_tests.sh matched the current master branch.